requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.8
psycopg2==2.9.9
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
